# api/openapi/gateway/v1/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "gateway/v1/gateway.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Gateway"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/miners": {
      "post": {
        "summary": "CreateMiner",
        "operationId": "Gateway_CreateMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/minersets": {
      "get": {
        "summary": "ListMinerSet",
        "operationId": "Gateway_ListMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMinerSetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "post": {
        "summary": "CreateMinerSet",
        "operationId": "Gateway_CreateMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/version": {
      "get": {
        "summary": "GetVersion",
        "operationId": "Gateway_GetVersion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Gateway"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetVersionResponse": {
      "type": "object",
      "properties": {
        "GitVersion": {
          "type": "string"
        },
        "GitCommit": {
          "type": "string"
        },
        "GitTreeState": {
          "type": "string"
        },
        "BuildDate": {
          "type": "string"
        },
        "GoVersion": {
          "type": "string"
        },
        "Compiler": {
          "type": "string"
        },
        "Platform": {
          "type": "string"
        }
      }
    },
    "v1ListMinerSetResponse": {
      "type": "object",
      "properties": {
        "totalCount": {
          "type": "string",
          "format": "int64"
        },
        "MinerSets": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1MinerSet"
          }
        }
      }
    },
    "v1MinerSet": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "replicas": {
          "type": "integer",
          "format": "int32"
        },
        "displayName": {
          "type": "string"
        },
        "deletePolicy": {
          "type": "string"
        },
        "MinerTemplate": {
          "$ref": "#/definitions/v1MinerTemplate"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1MinerTemplate": {
      "type": "object",
      "properties": {
        "minerType": {
          "type": "string"
        },
        "chainName": {
          "type": "string"
        },
        "dataDir": {
          "type": "string"
        }
      }
    }
  }
}